Extract `getSizeFromChildren` into a helper
